:newspaper: :hammer: Remove direct wandb dependency
